Rename mail component
