+++ src/main/java/pl/newsler/components/emaillabs/IMailRepository.java
@@ -1,4 +1,4 @@
-package pl.newsler.components.mail;
+package pl.newsler.components.emaillabs;
 
 public interface IMailRepository extends MailRepositoryJpa {
 }
