+++ src/main/java/pl/newsler/components/emaillabs/MailRepository.java
@@ -1,0 +1,4 @@
+package pl.newsler.components.emaillabs;
+
+public interface MailRepository {
+}
