+++ src/main/java/pl/newsler/components/emaillabs/MailRepositoryJpa.java
@@ -1,4 +1,4 @@
-package pl.newsler.components.mail;
+package pl.newsler.components.emaillabs;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.newsler.commons.models.NLId;
