+++ src/main/java/pl/newsler/components/emaillabs/NLUserMail.java
@@ -1,4 +1,4 @@
-package pl.newsler.components.mail;
+package pl.newsler.components.emaillabs;
 
 
 import lombok.AccessLevel;
