--- src/main/java/pl/newsler/components/mail/MailRepository.java
@@ -1,4 +1,0 @@
-package pl.newsler.components.mail;
-
-public interface MailRepository {
-}
